Time: 484 ms (30.38%), Space: 96.2 MB (46.34%) - LeetHub

+++ 1792-maximum-average-pass-ratio/1792-maximum-average-pass-ratio.java
@@ -1,8 +1,8 @@
 class Solution {
     public double maxAverageRatio(int[][] classes, int extraStudents) {
         Queue<double[]> passRatios = new PriorityQueue<>((a, b) -> Double.compare(
-                ((double) (b[0] + 1d) / (b[1] + 1d) - (double) b[0] / b[1]),
-                ((double) (a[0] + 1d) / (a[1] + 1d) - (double) a[0] / a[1])));
+                ((b[0] + 1d) / (b[1] + 1d) - (double) b[0] / b[1]),
+                ((a[0] + 1d) / (a[1] + 1d) - (double) a[0] / a[1])));
         for (int[] c : classes)
             passRatios.offer(new double[] { c[0], c[1] });
         while (extraStudents-- > 0) {
